+++ graphein/protein/utils.py
@@ -7,8 +7,11 @@ import logging
 # Project Website: https://github.com/a-r-j/graphein
 # Code Repository: https://github.com/a-r-j/graphein
 import os
+from functools import lru_cache
 from pathlib import Path
+from shutil import which
 from typing import Any, Dict, List, Tuple, Union
+from urllib.request import urlopen
 
 import networkx as nx
 import numpy as np
@@ -32,15 +35,38 @@ class ProteinGraphConfigurationError(Exception):
         return self.message
 
 
+@lru_cache()
+def get_obsolete_mapping() -> Dict[str, str]:
+    """Returns a dictionary mapping obsolete PDB codes to their replacement.
+
+    :return: Dictionary mapping obsolete PDB codes to their replacement.
+    :rtype: Dictionary[str, str]
+    """
+    obs_dict: Dict[str, str] = {}
+
+    response = urlopen("ftp://ftp.wwpdb.org/pub/pdb/data/status/obsolete.dat")
+    for line in response:
+        entry = line.split()
+        if len(entry) == 4:
+            obs_dict[entry[2].lower().decode("utf-8")] = (
+                entry[3].lower().decode("utf-8")
+            )
+    return obs_dict
+
+
 def download_pdb(config, pdb_code: str) -> Path:
     """
     Download PDB structure from PDB.
+
+    If no structure is found, we perform a lookup against the record of
+    obsolete PDB codes (ftp://ftp.wwpdb.org/pub/pdb/data/status/obsolete.dat)
 
     :param pdb_code: 4 character PDB accession code.
     :type pdb_code: str
     :return: returns filepath to downloaded structure.
     :rtype: str
     """
+    pdb_code = pdb_code.lower()
     if not config.pdb_dir:
         config.pdb_dir = Path("/tmp/")
 
@@ -49,6 +75,27 @@ def download_pdb(config, pdb_code: str) -> Path:
     pdbl.retrieve_pdb_file(
         pdb_code, pdir=config.pdb_dir, overwrite=True, file_format="pdb"
     )
+    # If file not downloaded, check for obsolescence
+    if not os.path.exists(config.pdb_dir / f"pdb{pdb_code}.ent"):
+        obs_map = get_obsolete_mapping()
+        try:
+            new_pdb = obs_map[pdb_code.lower()].lower()
+            log.info(
+                f"PDB file {pdb_code} not found. It is likely obsolete. \
+                     Trying its replacement: {new_pdb} instead."
+            )
+            pdb_code = new_pdb
+            pdbl.retrieve_pdb_file(
+                pdb_code,
+                pdir=config.pdb_dir,
+                overwrite=True,
+                file_format="pdb",
+            )
+        except KeyError:
+            log.error(
+                f"PDB file {pdb_code} not found and no replacement \
+                      structure found in obsolete lookup."
+            )
     # Rename file to .pdb from .ent
     os.rename(
         config.pdb_dir / f"pdb{pdb_code}.ent",
@@ -119,43 +166,67 @@ def compute_rgroup_dataframe(pdb_df: pd.DataFrame) -> pd.DataFrame:
 
 def download_alphafold_structure(
     uniprot_id: str,
+    version: int = 2,
     out_dir: str = ".",
+    rename: bool = True,
     pdb: bool = True,
     mmcif: bool = False,
     aligned_score: bool = True,
 ) -> Union[str, Tuple[str, str]]:
-    BASE_URL = "https://alphafold.ebi.ac.uk/files/"
     """
     Downloads a structure from the Alphafold EBI database (https://alphafold.ebi.ac.uk/files/").
 
     :param uniprot_id: UniProt ID of desired protein.
     :type uniprot_id: str
-    :param out_dir: String specifying desired output location. Default is pwd.
+    :param version: Version of the structure to download
+    :type version: int
+    :param out_dir: string specifying desired output location. Default is pwd.
     :type out_dir: str
-    :param mmcif: Bool specifying whether to download ``MMCiF`` or ``PDB``. Default is ``False`` (downloads pdb).
+    :param rename: boolean specifying whether to rename the output file to ``$uniprot_id.pdb``. Default is ``True``.
+    :type rename: bool
+    :param pdb: boolean specifying whether to download the PDB file. Default is ``True``.
+    :type pdb: bool
+    :param mmcif: Bool specifying whether to download MMCiF or PDB. Default is false (downloads pdb)
     :type mmcif: bool
     :param retrieve_aligned_score: Bool specifying whether or not to download score alignment json.
     :type retrieve_aligned_score: bool
     :return: path to output. Tuple if several outputs specified.
     :rtype: Union[str, Tuple[str, str]]
     """
+    BASE_URL = "https://alphafold.ebi.ac.uk/files/"
+    uniprot_id = uniprot_id.upper()
+
     if not mmcif and not pdb:
         raise ValueError("Must specify either mmcif or pdb.")
     if mmcif:
-        query_url = f"{BASE_URL}AF-{uniprot_id}F1-model_v1.cif"
+        query_url = f"{BASE_URL}AF-{uniprot_id}-F1-model_v{version}.cif"
     if pdb:
-        query_url = f"{BASE_URL}AF-{uniprot_id}-F1-model_v1.pdb"
-
+        query_url = f"{BASE_URL}AF-{uniprot_id}-F1-model_v{version}.pdb"
     structure_filename = wget.download(query_url, out=out_dir)
 
+    if rename:
+        extension = ".pdb" if pdb else ".cif"
+        os.rename(
+            structure_filename, Path(out_dir) / f"{uniprot_id}{extension}"
+        )
+        structure_filename = str(
+            (Path(out_dir) / f"{uniprot_id}{extension}").resolve()
+        )
+
+    log.info(f"Downloaded AlphaFold PDB file for: {uniprot_id}")
     if aligned_score:
         score_query = (
             BASE_URL
             + "AF-"
             + uniprot_id
-            + "-F1-predicted_aligned_error_v1.json"
+            + f"-F1-predicted_aligned_error_v{version}.json"
         )
         score_filename = wget.download(score_query, out=out_dir)
+        if rename:
+            os.rename(score_filename, Path(out_dir) / f"{uniprot_id}.json")
+            score_filename = str(
+                (Path(out_dir) / f"{uniprot_id}.json").resolve()
+            )
         return structure_filename, score_filename
 
     return structure_filename
@@ -229,3 +300,16 @@ def save_rgroup_df_to_pdb(g: nx.Graph, path: str, gz: bool = False):
     ppd.df["ATOM"] = g.graph["rgroup_df"]
     ppd.to_pdb(path=path, records=None, gz=gz, append_newline=True)
     log.info(f"Successfully saved rgroup data to {path}")
+
+
+def is_tool(name: str) -> bool:
+    """Checks whether ``name`` is on PATH and is marked as an executable.
+    
+    Source: https://stackoverflow.com/questions/11210104/check-if-a-program-exists-from-a-python-script
+
+    :param name: Name of program to check for execution ability.
+    :type name: str
+    :return: Whether ``name`` is on PATH and is marked as an executable.
+    :rtype: bool
+    """
+    return which(name) is not None