+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-# 1.2.2 - Unreleased
+### 1.3.1 - UNRELEASED
+* [Feature] - #154 adds a way of checking that DSSP is executable before trying to use it. #154
+
+### 1.3.0 - 5/4/22
+
 * [Feature] - #141 adds edge construction based on sequence distance.
 * [Feature] - #143 adds equality and isomorphism testing functions between graphs, nodes and edges ([#142](https://github.com/a-r-j/graphein/issues/142))
+* [Feature] - #144 adds support for chain-level and secondary structure-level graphs with associated visualisation tools and tutorial. Resolves [#128](https://github.com/a-r-j/graphein/issues/128)
+* [Feature] - #144 adds support for chord diagram visualisations.
+* [Feature] - #144 adds support for automagically downloading new PDB files for obsolete structures.
+* [Feature] - #150 adds support for hydrogen bond donor and acceptor counts node features. #145
+* [Misc] - #144 makes visualisation functions accessible in the `graphein.protein` namespace. #138
+* [Bugfix] - #147 fixes error in `add_distance_threshold` introduced in v1.2.1 that would prevent the edges being added to the graph. [#146](https://github.com/a-r-j/graphein/issues/146) 
+* [Bugfix] - #149 fixes a bug in `add_beta_carbon_vector` that would cause coordinates to be extracted for multiple positions if the residue has an altloc. Resolves [#148](https://github.com/a-r-j/graphein/issues/148)
 
-
-### 1.2.1 - 16/3/21
+### 1.2.1 - 16/3/22
 
 * [Feature] - #124 adds support for vector features associated protein protein geometry. #120 #122
 * [Feature] - #124 adds visualisation of vector features in 3D graph plots.